--- contexts/chat.provider.web.tsx
@@ -1,11 +1,0 @@
-import { PropsWithChildren } from "react";
-
-export const ChatProvider: React.FC<PropsWithChildren> = ({
-  children,
-}) => {
-  return (
-    <>
-      {children}
-    </>
-  );
-};
