+++ hooks/index.ts
@@ -1,6 +1,5 @@
 import useBreakpoints from "./use-breakpoints";
 import { useStorageState } from "./use-storage-state";
 import useStreamTheme from "./use-stream-theme";
-import { useChatHook } from "./use-chat";
 
-export { useBreakpoints, useStorageState, useChatHook, useStreamTheme };
+export { useBreakpoints, useStorageState, useStreamTheme };
